setting up the entrypoint
