+++ .gitignore
@@ -3,6 +3,5 @@ dist
 coverage
 .nyc_output
 node_modules
-/starter/**/*.js
-/starter/**/*.js.map
-
+**/*.js
+**/*.js.map
