+++ package.json
@@ -1,15 +1,19 @@
 {
   "name": "create-fullstacked",
-  "version": "0.0.15",
+  "version": "0.0.0",
   "description": "Create a FullStacked Project",
-  "bin": "index.js",
+  "bin": {
+    "create-fullstacked": "index.js"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/CPLepage/create-fullstacked.git"
   },
   "scripts": {
-    "start": "npx fullstacked watch --src=starter",
-    "test": "npx fullstacked test --src=starter"
+    "build": "npx esbuild index.ts --outfile=index.js --format=cjs",
+    "pretest": "npm run build && npx esbuild test.ts --outfile=test.js --format=cjs",
+    "test": "node test",
+    "prepublishOnly": "npm run build"
   },
   "author": "CP Lepage",
   "license": "MIT",
@@ -18,9 +22,11 @@
   },
   "homepage": "https://fullstacked.org",
   "devDependencies": {
-    "fullstacked": "^0.6.1"
+    "fullstacked": "^0.8.0-alpha2"
   },
   "dependencies": {
     "fs-extra": "^10.1.0"
-  }
+  },
+  "main": "index.js",
+  "keywords": []
 }
