+++ templates/default/server/index.ts
@@ -1,0 +1,7 @@
+import Server from "fullstacked/server";
+
+const server = new Server();
+
+server.start();
+
+export default server;
