+++ templates/default/tests/e2e.test.ts
@@ -1,3 +1,4 @@
+import {describe, it, before, after} from "mocha";
 import Helper from "fullstacked/tests/e2e/Helper";
 import {equal} from "assert";
 import * as path from "path";
