+++ templates/pwa/webapp/service-worker.ts
