+++ test.ts
@@ -1,0 +1,5 @@
+import * as path from "path";
+import * as fs from "fs";
+
+const templates = fs.readdirSync(path.resolve(__dirname, "templates"));
+
